+++ src/lib/sync/poly-iterable.ts
@@ -300,6 +300,40 @@ export default class PolySyncIterable<T> implements Iterable<T> {
   }
 
   /**
+   * Return an iteration of the pairs resulting of calling `func(element)` for every element in `this` and using it as
+   * the first element of the pair (the *key*) and preserving the second (the *value*).
+   *
+   * @remarks
+   * This method is only available for iterations of pairs.
+   *
+   * @typeParam U - The return type of `func` and the generic type of the resulting iterable
+   * @param func - A function that takes an element of `this` and returns something else
+   * @returns A new {@link PolySyncIterable} that yields the results of calling `func(element)`
+   * for every element of `this` and using it to replace the keys
+   */
+  mapKeys<K1, K2, V> (this: PolySyncIterable<[K1, V]>, func: IndexedMapping<[K1, V], K2>): PolySyncIterable<[K2, V]> {
+    asserts.isFunction(func)
+    return new PolySyncIterable(mapGen(this.#iterable, ([k, v], index) => [func([k, v], index), v]))
+  }
+
+  /**
+   * Return an iteration of the pairs resulting of calling `func(element)` for every element in `this` and using it as
+   * the second element of the pair (the *value*) and preserving the first (the *key*).
+   *
+   * @remarks
+   * This method is only available for iterations of pairs.
+   *
+   * @typeParam U - The return type of `func` and the generic type of the resulting iterable
+   * @param func - A function that takes an element of `this` and returns something else
+   * @returns A new {@link PolySyncIterable} that yields the results of calling `func(element)`
+   * for every element of `this` and using it to replace the values
+   */
+  mapValues<K, V1, V2> (this: PolySyncIterable<[K, V1]>, func: IndexedMapping<[K, V1], V2>): PolySyncIterable<[K, V2]> {
+    asserts.isFunction(func)
+    return new PolySyncIterable(mapGen(this.#iterable, ([k, v], index) => [k, func([k, v], index)]))
+  }
+
+  /**
    * Return an iteration of the same elements as `this` after calling `func(element)` for all elements.
    *
    * @typeParam U - The return type of `func`
@@ -530,7 +564,7 @@ export default class PolySyncIterable<T> implements Iterable<T> {
    *
    * @returns An object composed of the entries yielded by this iterable.
    */
-  toObject<K extends PropertyKey, V> (this: PolySyncIterable<readonly [K, V]>): Record<K, V> {
+  toObject<K extends PropertyKey, V> (this: PolySyncIterable<[K, V] | readonly [K, V]>): Record<K, V> {
     const object = {} as Record<K, V>
 
     for (const [key, value] of this.#iterable) {
@@ -549,7 +583,7 @@ export default class PolySyncIterable<T> implements Iterable<T> {
    *
    * @returns A `Map` composed of the entries yielded by this iterable.
    */
-  toMap<K, V> (this: PolySyncIterable<readonly [K, V]>): Map<K, V> {
+  toMap<K, V> (this: PolySyncIterable<[K, V] | readonly [K, V]>): Map<K, V> {
     const map = new Map<K, V>()
 
     for (const [key, value] of this.#iterable) {