+++ CHANGELOG.md
@@ -6,6 +6,18 @@ This file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## UNRELEASED
+
+### Added:
+  - Added `mapKey` method to both iterables
+  - Added `mapValue` method to both iterables
+
+### Changed:
+  - Improved types of `toObject` and `toMap` methods
+  - Add an `options` argument to `PolyAsyncIterable#complete`
+  - Add an `options` argument to `PolyAsyncIterable#join`
+
+
 ## [2.3.0] - 2023-03-16
 
 ### Added: